move GMMClusterImpl to its own module
